Decouple JH configuration from Spawner definition
* Extract KubeFormSpawner class definition to a new file,
  named `spawner.py`
* Rename `kubeform_spawner.py` to `jupyterhub_config.py`,
  as it now contains the `JupyterHub` class configuration
* Pass needed environmental variables via `KubeSpawner.environment`
  attribute instead of calling `os.environ.get()`, inside `spawner.py`
* Fix imports and Python styling
